Calculates total consumption by aggregating all samples from all devices

+++ elis-energy-service/src/main/java/se/mah/elis/external/energy/beans/EnergySummaryBean.java
@@ -10,5 +10,5 @@ public class EnergySummaryBean {
 	public String deviceId;
 	
 	@XmlElement
-	public float kwh;
+	public double kwh;
 }
